Statistics on an independent tab (with info, jobs, participants), not in participants. Order the tabs like this: INFO / JOBS / PARTICIPANTS / STATISTICS
