+++ resources/components/Participants/ParticipationStats/ParticipationStats.test.js
