+++ resources/components/Participants/ParticipationStats/index.js
